convert from js to ts
