+++ package.json
@@ -13,8 +13,8 @@
     "packages/*"
   ],
   "scripts": {
-    "start": "node -r esm index.js",
-    "watch": "nodemon -r esm index.js",
+    "start": "ts-node -r tsconfig-paths/register ./index.ts",
+    "watch": "nodemon --exec ts-node index.ts",
     "build": "yarn webpack --config packages/client/webpack.config.js --mode production"
   },
   "cacheDirectories": [
@@ -31,16 +31,24 @@
     "esm": "^3.2.25",
     "express": "^4.17.1",
     "form-data": "^3.0.0",
+    "ts-node": "^8.10.2",
     "webpack": "^4.43.0",
     "webpack-dev-middleware": "^3.7.2"
   },
   "devDependencies": {
+    "@babel/preset-typescript": "^7.10.4",
+    "@types/btoa": "^1.2.3",
     "@types/classnames": "^2.2.10",
+    "@types/cors": "^2.8.7",
     "@types/css-modules": "^1.0.0",
+    "@types/dedent": "^0.7.0",
+    "@types/express": "^4.17.7",
     "@types/node": "^14.0.18",
     "@types/react": "^16.9.41",
     "@types/react-dom": "^16.9.8",
     "@types/react-router-dom": "^5.1.5",
+    "@types/sequelize": "^4.28.9",
+    "@types/webpack-dev-middleware": "^3.7.1",
     "@types/webpack-env": "^1.15.2",
     "awesome-typescript-loader": "^5.2.1",
     "babel-eslint": "^10.1.0",
@@ -58,6 +66,7 @@
     "prettier": "^2.0.5",
     "source-map-loader": "^1.0.1",
     "ts-loader": "^7.0.5",
+    "tsconfig-paths": "^3.9.0",
     "typescript": "^3.9.6"
   }
 }
