+++ packages/bot/commands/index.ts
