+++ packages/bot/commands/sticker.ts
@@ -1,0 +1,52 @@
+import { MessageEmbed } from 'discord.js';
+import fs from 'fs';
+import path from 'path';
+import Command from '../types/command';
+
+const files = fs.readdirSync(path.join(__dirname, 'stickers'));
+const stickers = files.map((file) => {
+  return {
+    name: file.split('.')[0],
+    path: path.join(__dirname, 'stickers', file),
+  };
+});
+
+export default {
+  config: {
+    enabled: true,
+    permissionLevel: 0,
+    aliases: ['stamp', 'st', 'stickers'],
+    guildOnly: true,
+    ownerOnly: false,
+  },
+
+  help: {
+    name: 'sticker',
+    description: 'Sends a sticker',
+    usage: '\n{prefix} [sticker name]\n{prefix} list',
+  },
+
+  run: async (client, message, _, params) => {
+    const [command] = params;
+
+    if (!command || command.toLowerCase() === 'list') {
+      message.channel.send(
+        `You can view a list of stickers at https://${client.options.host}/stickers`
+      );
+
+      return;
+    }
+
+    const sticker = stickers.find((s) => s.name.toLowerCase() === command.toLowerCase());
+    if (!sticker) {
+      message.channel.send(`Could nyot find this sticker..`);
+      return;
+    }
+
+    const embed = new MessageEmbed()
+      .attachFiles([sticker.path])
+      .setImage(`attachment://${sticker.name}.png`);
+
+    message.channel.send({ embed });
+  },
+} as Command;
