+++ packages/bot/index.ts
